// apps/posts/src/infrastructure/prisma/generated/models/OutboxEvent.ts
export type OutboxEventModel = runtime.Types.Result.DefaultSelection<Prisma.$OutboxEventPayload>
// ...
export type $OutboxEventPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "OutboxEvent"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    eventType: string
    aggregateType: string
    aggregateId: string
    payload: runtime.JsonValue
    status: $Enums.OutboxStatus
    attempts: number
    availableAt: Date
    createdAt: Date
    publishedAt: Date | null
    lastError: string | null
  }, ExtArgs["result"]["outboxEvent"]>
  composites: {}
}